Notification and ajax added

+++ controller/HomeController.php
@@ -7,6 +7,7 @@ class HomeController extends Controller
     public function __construct()
     {
         $this->UserModel =  new UserModel();
+        $this->BASE_URL = "http://localhost/LSS/index.php";
     }
 
     public function index()
@@ -37,7 +38,8 @@ class HomeController extends Controller
     public function home($msg = "", $msgs = "")
     {
         if (!isset($_SESSION['loggedin'])) {
-            return $this->login($msg);
+            header("Location: " . $this->BASE_URL . "?act=login");
+            exit;
         }
         $result = $this->UserModel->getUser($_SESSION['useremail']);
 
@@ -47,34 +49,41 @@ class HomeController extends Controller
         $data['title'] = 'Home page';
         $data['msg'] = $msg;
         $data['msgS'] = $msgs;
-        $this->view('nav', $data);
+        // $this->view('nav', $data);
         $this->view('home', $data);
-        $this->view('footer');
+        // $this->view('footer');
     }
 
     public function delete()
     {
+        if (!isset($_SESSION['loggedin'])) {
+            header("Location: " . $this->BASE_URL . "?act=login");
+            exit;
+        }
         if (true == isset($_POST['userId'])) {
-            $id = trim($_POST["userId"]);
+            $id = $this->cleanData($_POST["userId"]);
             $this->UserModel->delete($id);
             unset($_SESSION['loggedin']);
             unset($_SESSION['useremail']);
-            $msg = "User deleted";
-            return $this->home($msg);
+            echo json_encode(array("result" => "deleted"));
         }
     }
 
     public function update()
     {
+        if (!isset($_SESSION['loggedin'])) {
+            header("Location: " . $this->BASE_URL . "?act=login");
+            exit;
+        }
 
         if (true == isset($_POST['firstname'])) {
-            $id = trim($_POST["id"]);
-            $firstname = trim($_POST["firstname"]);
-            $lastname = trim($_POST["lastname"]);
-            $username = trim($_POST["username"]);
-            $email = trim($_POST["email"]);
-            $phone = trim($_POST["phone"]);
-            $password = trim($_POST["password"]);
+            $id = $this->cleanData($_POST["id"]);
+            $firstname = $this->cleanData($_POST["firstname"]);
+            $lastname = $this->cleanData($_POST["lastname"]);
+            $username = $this->cleanData($_POST["username"]);
+            $email = $this->cleanData($_POST["email"]);
+            $phone = $this->cleanData($_POST["phone"]);
+            $password = $this->cleanData($_POST["password"]);
             $userdata = array(
                 $firstname,
                 $lastname,
@@ -86,57 +95,68 @@ class HomeController extends Controller
 
             if ($_SESSION['useremail'] != $email  and $this->UserModel->alreadyExist($email, "forUpdate")) {
                 $msgs = 'email already exisist';
-                return $this->home("", $msgs);
+                echo json_encode(array("is_update" => false));
+                return;
             }
             $_SESSION['useremail'] = $email;
             $this->UserModel->update($userdata, $id);
             $msg = 'updated success';
-            return $this->home($msg);
+            echo json_encode(array("is_update" => true));
+            return;
         }
     }
 
     public function logout()
     {
+        if (!isset($_SESSION['loggedin'])) {
+            header("Location: " . $this->BASE_URL . "?act=login");
+            exit;
+        }
         $_SESSION['loggedin'] = false;
         $_SESSION['useremail'] = "";
         unset($_SESSION['loggedin']);
         unset($_SESSION['useremail']);
         $data['title'] = 'Login';
-        $this->view('nav', $data);
-        $this->view('login', $data);
-        $this->view('footer');
+        echo json_encode(array("k" => "logout", "data" => $data));
+        // $this->view('login', $data);
     }
 
 
     public function login($msg = "")
     {
+        if (isset($_SESSION['loggedin'])) {
+            header("Location: " . $this->BASE_URL . "");
+            exit;
+        }
+        $data['msgS'] = $msg;
         if (true == isset($_POST['email'])) {
-            $email = trim($_POST["email"]);
-            $password = trim($_POST["password"]);
+            $email = $this->cleanData($_POST["email"]);
+            $password =  $this->cleanData($_POST["password"]);
             if ($this->UserModel->login($email, $password)) {
                 $_SESSION['loggedin'] = true;
                 $_SESSION['useremail'] = $email;
-                return $this->home();
+                header("Location: " . $this->BASE_URL . "");
             } else {
                 $data['msgS'] = 'email or password is incorrect';
             }
         }
         $data['title'] = 'Login';
-        $data['msgS'] = $msg;
-        $this->view('nav', $data);
         $this->view('login', $data);
-        $this->view('footer');
     }
 
     public function signUp()
     {
+        if (isset($_SESSION['loggedin'])) {
+            header("Location: " . $this->BASE_URL . "");
+            exit;
+        }
         if (true == isset($_POST['firstname'])) {
-            $firstname = trim($_POST["firstname"]);
-            $lastname = trim($_POST["lastname"]);
-            $username = trim($_POST["username"]);
-            $email = trim($_POST["email"]);
-            $phone = trim($_POST["phone"]);
-            $password = trim($_POST["password"]);
+            $firstname = $this->cleanData($_POST["firstname"]);
+            $lastname = $this->cleanData($_POST["lastname"]);
+            $username = $this->cleanData($_POST["username"]);
+            $email = $this->cleanData($_POST["email"]);
+            $phone = $this->cleanData($_POST["phone"]);
+            $password = $this->cleanData($_POST["password"]);
             $userdata = array(
                 $firstname,
                 $lastname,
@@ -150,23 +170,25 @@ class HomeController extends Controller
                 $data['title'] = 'signUp';
                 $data['userData'] = $userdata;
                 $data['msg'] = "email already exisit";
-                $this->view('nav', $data);
                 $this->view('signup', $data);
-                $this->view('footer');
                 return;
             } else {
                 $this->UserModel->signUp($userdata);
                 $data['title'] = 'Login';
-                $data['msg'] = "Successfully Resgister Login Now..";
-                $this->view('nav', $data);
-                $this->view('login', $data);
-                $this->view('footer');
+                $_SESSION['loggedin'] = true;
+                $_SESSION['useremail'] = $email;
+                header("Location: " . $this->BASE_URL . "");
+                return;
             }
         } else {
             $data['title'] = 'signUp';
-            $this->view('nav', $data);
             $this->view('signup', $data);
-            $this->view('footer');
         }
+    }
+
+    public function cleanData($value)
+    {
+        $value = mysqli_real_escape_string($this->UserModel->conn, trim($value));
+        return $value;
     }
 }
